[ refactor ] 폴더 정리
