--- src/@components/JoinPage/UserProfilePage/ProfileImage/indes.tsx
@@ -1,5 +1,0 @@
-import { St } from "./style";
-
-export default function ProfileImage() {
-  return <St.ProfileImage></St.ProfileImage>;
-}
