+++ src/@components/JoinPage/UserProfilePage/ProfileImage/style.ts
@@ -2,4 +2,25 @@ import styled from "styled-components";
 
 export const St = {
   ProfileImage: styled.section``,
+
+  Image: styled.div`
+    width: 8rem;
+    height: 8rem;
+
+    position: relative;
+    margin: 3.2rem auto;
+    display: block;
+  `,
+
+  AddImage: styled.img``,
+
+  AddBtnWrapper: styled.label`
+    position: absolute;
+    top: 5.8rem;
+    left: 6rem;
+  `,
+
+  AddButton: styled.input`
+    display: none;
+  `,
 };
