--- src/@components/JoinPage/UserProfilePage/ProfileJender/indes.tsx
@@ -1,5 +1,0 @@
-import { St } from "./style";
-
-export default function ProfileImage() {
-  return <St.ProfileJender></St.ProfileJender>;
-}
