+++ src/@components/JoinPage/UserProfilePage/ProfileJender/index.tsx
@@ -1,0 +1,15 @@
+import { IcDownArrow } from "../../../../asset/icon";
+import { St } from "./style";
+
+export default function ProfileImage() {
+  return (
+    <St.ProfileJender>
+      <St.InputContainer>
+        <St.GenderInputForm id="gender" placeholder="성별을 체크해주세요" />
+        <St.Down>
+          <IcDownArrow />
+        </St.Down>
+      </St.InputContainer>
+    </St.ProfileJender>
+  );
+}
