--- src/@components/JoinPage/UserProfilePage/ProfileNickname/indes.tsx
@@ -1,5 +1,0 @@
-import { St } from "./style";
-
-export default function ProfileImage() {
-  return <St.ProfileNickname></St.ProfileNickname>;
-}
