+++ src/@components/JoinPage/UserProfilePage/ProfileNickname/index.tsx
@@ -1,0 +1,12 @@
+import { St } from "./style";
+
+export default function ProfileImage() {
+  return (
+    <St.ProfileNickname>
+      <St.InputContainer>
+        <St.NickNameInputForm id="nickname" placeholder="홍길동" />
+        <St.CheckBtn>중복확인</St.CheckBtn>
+      </St.InputContainer>
+    </St.ProfileNickname>
+  );
+}
